refactor(examples): replace react-scripts with parcel
the main reason is that react-scripts version here is quite outdated, and causing some duplications and useless packages in our lockfile/node_modules
