+++ examples/react/autocomplete/__mocks__/styleMock.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/no-commonjs
+module.exports = {};
