+++ examples/react/autocomplete/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>autocomplete with react-instantsearch</title>
+    <link rel="stylesheet" type="text/css" href="src/style.css" />
+  </head>
+
+  <body>
+    <h2>Autocomplete targeting multiple indices</h2>
+    <div id="autocomplete-with-multi-indices"></div>
+    <h2>Autocomplete of mentions</h2>
+    <div id="autocomplete-mentions"></div>
+
+    <script type="module" src="src/index.js"></script>
+  </body>
+</html>
