--- examples/react/autocomplete/public/index.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>autocomplete with react-instantsearch</title>
-  <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/style.css">
-</head>
-
-<body>
-  <h2>Autocomplete targetting multiple indices</h2>
-  <div id="autocomplete-with-multi-indices"></div>
-  <h2>Autocomplete of mentions</h2>
-  <div id="autocomplete-mentions"></div>
-</body>
-
-</html>
