+++ examples/react/autocomplete/src/__tests__/App.test.js
@@ -1,6 +1,10 @@
+/**
+ * @jest-environment jsdom
+ */
+
 import React from 'react';
-import MultiIndex from './App-Multi-Index';
-import Mentions from './App-Mentions';
+import MultiIndex from '../App-Multi-Index';
+import Mentions from '../App-Mentions';
 import renderer from 'react-test-renderer';
 
 jest.mock('antd/lib/mention', () => {
