+++ examples/react/autocomplete/src/style.css
@@ -1,0 +1,115 @@
+body {
+  font-family: Helvetica, sans-serif;
+  margin: 0;
+  padding: 1em;
+  box-sizing: border-box;
+}
+
+.highlighted {
+  background: yellow;
+}
+
+.autocomplete {
+  display: flex;
+  justify-content: space-between;
+  border: 1px solid red;
+}
+
+.content {
+  display: flex;
+  justify-content: space-between;
+}
+
+.block {
+  padding: 5px 10px;
+}
+
+.title {
+  border: 1px solid red;
+  display: flex;
+  justify-content: space-around;
+  padding: 5px 10px;
+}
+
+.disabled {
+  display: none;
+}
+
+.ais-MultiIndex__root {
+  width: 100%;
+}
+
+.title div {
+  width: 100%;
+  padding: 5px 10px;
+}
+
+.react-autosuggest__container {
+  position: relative;
+}
+
+.react-autosuggest__input {
+  width: 600px;
+  height: 30px;
+  padding: 10px 20px;
+  font-family: Helvetica, sans-serif;
+  font-weight: 300;
+  font-size: 16px;
+  border: 1px solid #aaa;
+  border-radius: 4px;
+}
+
+.react-autosuggest__input:focus {
+  outline: none;
+}
+
+.react-autosuggest__container--open .react-autosuggest__input {
+  border-bottom-left-radius: 0;
+  border-bottom-right-radius: 0;
+}
+
+.react-autosuggest__suggestions-container {
+  display: none;
+}
+
+.react-autosuggest__container--open .react-autosuggest__suggestions-container {
+  display: block;
+  position: absolute;
+  top: 51px;
+  width: 640px;
+  border: 1px solid #aaa;
+  background-color: #fff;
+  font-family: Helvetica, sans-serif;
+  font-weight: 300;
+  font-size: 16px;
+  border-bottom-left-radius: 4px;
+  border-bottom-right-radius: 4px;
+  z-index: 2;
+}
+
+.react-autosuggest__suggestions-list {
+  margin: 0;
+  padding: 0;
+  list-style-type: none;
+}
+
+.react-autosuggest__suggestion {
+  cursor: pointer;
+  padding: 10px 20px;
+}
+
+.react-autosuggest__suggestion--focused {
+  background-color: #ddd;
+}
+
+.react-autosuggest__section-title {
+  padding: 10px 0 0 10px;
+  font-size: 12px;
+  color: #777;
+  border-top: 1px dashed #ccc;
+}
+
+.react-autosuggest__section-container:first-child
+  .react-autosuggest__section-title {
+  border-top: 0;
+}
