--- examples/react/geo-search/public/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>GeoSearch with React InstantSearch</title>
-</head>
-
-<body>
-  <div id="root"></div>
-</body>
-
-</html>
