+++ examples/react/multi-index/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>multi-index search with react-instantsearch</title>
+  </head>
+  <body>
+    <div id="root"></div>
+
+    <script type="module" src="./src/index.js"></script>
+  </body>
+</html>
