+++ examples/react/multi-index/package.json
@@ -4,12 +4,16 @@
   "private": true,
   "license": "MIT",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "SKIP_PREFLIGHT_CHECK=true CI=true react-scripts test"
+    "start": "parcel index.html --port 3000",
+    "build": "parcel build index.html --public-url .",
+    "test": "jest --ci"
   },
   "devDependencies": {
-    "react-scripts": "3.1.2",
+    "@babel/core": "7.15.5",
+    "@parcel/core": "2.8.0",
+    "@parcel/packager-raw-url": "2.8.0",
+    "@parcel/transformer-webmanifest": "2.8.0",
+    "parcel": "2.8.0",
     "react-test-renderer": "17.0.2"
   },
   "dependencies": {
