+++ examples/react/multi-index/src/__tests__/App.test.js
@@ -1,8 +1,12 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import App from './App';
+/**
+ * @jest-environment jsdom
+ */
 
-describe('geo-search recipe', () => {
+import React from 'react';
+import App from '../App';
+import renderer from 'react-test-renderer';
+
+describe('Multi index recipe', () => {
   it('App renders without crashing', () => {
     const component = renderer.create(<App />);
