+++ examples/react/react-router-v3/index.html
@@ -3,9 +3,11 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>multi-index search with react-instantsearch</title>
+    <title>react-router feat react-instantsearch</title>
   </head>
   <body>
     <div id="root"></div>
+
+    <script type="module" src="src/index.js"></script>
   </body>
 </html>
