+++ examples/react/react-router/index.html
@@ -7,5 +7,7 @@
   </head>
   <body>
     <div id="root"></div>
+
+    <script type="module" src="src/index.js"></script>
   </body>
 </html>
