--- examples/react/react-router/public/index.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>react-router feat react-instantsearch</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
